started abstracting command to command_base
