+++ nextcord/ext/abc/_types.py
@@ -49,3 +49,4 @@ CogT = TypeVar('CogT', bound='Cog')
 CommandT = TypeVar('CommandT', bound='CommandBase')
 ContextT = TypeVar('ContextT', bound='ContextBase')
 CoroFuncT = TypeVar('CoroFuncT', bound='CoroFunc')
+HookT = TypeVar('HookT', bound='Hook')
