+++ nextcord/ext/commands/core.py
@@ -124,63 +124,12 @@ CogT = TypeVar('CogT', bound='Cog')
 CommandT = TypeVar('CommandT', bound='Command')
 ContextT = TypeVar('ContextT', bound='Context')
 GroupT = TypeVar('GroupT', bound='Group')
-HookT = TypeVar('HookT', bound='Hook')
 ErrorT = TypeVar('ErrorT', bound='Error')
 
 if TYPE_CHECKING:
     P = ParamSpec('P')
 else:
     P = TypeVar('P')
-
-
-def unwrap_function(function: Callable[..., Any]) -> Callable[..., Any]:
-    partial = functools.partial
-    while True:
-        if hasattr(function, '__wrapped__'):
-            function = function.__wrapped__
-        elif isinstance(function, partial):
-            function = function.func
-        else:
-            return function
-
-
-def get_signature_parameters(function: Callable[..., Any], globalns: Dict[str, Any]) -> Dict[str, inspect.Parameter]:
-    signature = inspect.signature(function)
-    params = {}
-    cache: Dict[str, Any] = {}
-    eval_annotation = nextcord.utils.evaluate_annotation
-    for name, parameter in signature.parameters.items():
-        annotation = parameter.annotation
-        if annotation is parameter.empty:
-            params[name] = parameter
-            continue
-        if annotation is None:
-            params[name] = parameter.replace(annotation=type(None))
-            continue
-
-        annotation = eval_annotation(annotation, globalns, globalns, cache)
-        if annotation is Greedy:
-            raise TypeError(
-                'Unparameterized Greedy[...] is disallowed in signature.')
-
-        params[name] = parameter.replace(annotation=annotation)
-
-    return params
-
-
-def wrap_callback(coro):
-    @functools.wraps(coro)
-    async def wrapped(*args, **kwargs):
-        try:
-            ret = await coro(*args, **kwargs)
-        except CommandError:
-            raise
-        except asyncio.CancelledError:
-            return
-        except Exception as exc:
-            raise CommandInvokeError(exc) from exc
-        return ret
-    return wrapped
 
 
 def hooked_wrapped_callback(command, ctx, coro):
@@ -226,7 +175,7 @@ class _CaseInsensitiveDict(dict):
         super().__setitem__(k.casefold(), v)
 
 
-class Command(CommandBase, Generic[CogT, P, T]):
+class Command(CommandBase):
     r"""A class that implements the protocol for a bot text command.
 
     These are not created manually, instead they are created via the
@@ -323,243 +272,26 @@ class Command(CommandBase, Generic[CogT, P, T]):
         Callable[Concatenate[CogT, ContextT, P], Coro[T]],
         Callable[Concatenate[ContextT, P], Coro[T]],
     ], **kwargs: Any):
-        if not asyncio.iscoroutinefunction(func):
-            raise TypeError('Callback must be a coroutine.')
+        super().__init__(func, **kwargs)
 
-        name = kwargs.get('name') or func.__name__
-        if not isinstance(name, str):
-            raise TypeError('Name of a command must be a string.')
-        self.name: str = name
-
-        self.callback = func
-        self.enabled: bool = kwargs.get('enabled', True)
-
-        help_doc = kwargs.get('help')
-        if help_doc is not None:
-            help_doc = inspect.cleandoc(help_doc)
-        else:
-            help_doc = inspect.getdoc(func)
-            if isinstance(help_doc, bytes):
-                help_doc = help_doc.decode('utf-8')
-
-        self.help: Optional[str] = help_doc
-
-        self.brief: Optional[str] = kwargs.get('brief')
-        self.usage: Optional[str] = kwargs.get('usage')
+    def set_fields(self, func: Union[
+        Callable[Concatenate[CogT, ContextT, P], Coro[T]],
+        Callable[Concatenate[ContextT, P], Coro[T]],
+    ], **kwargs: Any):
+        super().set_fields(func, **kwargs)
         self.rest_is_raw: bool = kwargs.get('rest_is_raw', False)
         self.aliases: Union[List[str], Tuple[str]] = kwargs.get('aliases', [])
-        self.extras: Dict[str, Any] = kwargs.get('extras', {})
-
         if not isinstance(self.aliases, (list, tuple)):
-            raise TypeError(
-                "Aliases of a command must be a list or a tuple of strings.")
-
-        self.description: str = inspect.cleandoc(kwargs.get('description', ''))
-        self.hidden: bool = kwargs.get('hidden', False)
-
-        try:
-            checks = func.__commands_checks__
-            checks.reverse()
-        except AttributeError:
-            checks = kwargs.get('checks', [])
-
-        self.checks: List[Check] = checks
-
-        try:
-            cooldown = func.__commands_cooldown__
-        except AttributeError:
-            cooldown = kwargs.get('cooldown')
-
-        if cooldown is None:
-            buckets = CooldownMapping(cooldown, BucketType.default)
-        elif isinstance(cooldown, CooldownMapping):
-            buckets = cooldown
-        else:
-            raise TypeError(
-                "Cooldown must be a an instance of CooldownMapping or None.")
-        self._buckets: CooldownMapping = buckets
-
-        try:
-            max_concurrency = func.__commands_max_concurrency__
-        except AttributeError:
-            max_concurrency = kwargs.get('max_concurrency')
-
-        self._max_concurrency: Optional[MaxConcurrency] = max_concurrency
+            raise TypeError("Aliases of a command must be a list or a tuple of strings.")
 
         self.require_var_positional: bool = kwargs.get(
             'require_var_positional', False)
         self.ignore_extra: bool = kwargs.get('ignore_extra', True)
-        self.cooldown_after_parsing: bool = kwargs.get(
-            'cooldown_after_parsing', False)
-        self.cog: Optional[CogT] = None
 
         # bandaid for the fact that sometimes parent can be the bot instance
         parent = kwargs.get('parent')
         self.parent: Optional[GroupMixin] = parent if isinstance(
             parent, CommandBase) else None  # type: ignore
-
-        self._before_invoke: Optional[Hook] = None
-        try:
-            before_invoke = func.__before_invoke__
-        except AttributeError:
-            pass
-        else:
-            self.before_invoke(before_invoke)
-
-        self._after_invoke: Optional[Hook] = None
-        try:
-            after_invoke = func.__after_invoke__
-        except AttributeError:
-            pass
-        else:
-            self.after_invoke(after_invoke)
-
-    @property
-    def callback(self) -> Union[
-        Callable[Concatenate[CogT, Context, P], Coro[T]],
-        Callable[Concatenate[Context, P], Coro[T]],
-    ]:
-        return self._callback
-
-    @callback.setter
-    def callback(self, function: Union[
-        Callable[Concatenate[CogT, Context, P], Coro[T]],
-        Callable[Concatenate[Context, P], Coro[T]],
-    ]) -> None:
-        self._callback = function
-        unwrap = unwrap_function(function)
-        self.module = unwrap.__module__
-
-        try:
-            globalns = unwrap.__globals__
-        except AttributeError:
-            globalns = {}
-
-        self.params = get_signature_parameters(function, globalns)
-
-    def add_check(self, func: Check) -> None:
-        """Adds a check to the command.
-
-        This is the non-decorator interface to :func:`.check`.
-
-        .. versionadded:: 1.3
-
-        Parameters
-        -----------
-        func
-            The function that will be used as a check.
-        """
-
-        self.checks.append(func)
-
-    def remove_check(self, func: Check) -> None:
-        """Removes a check from the command.
-
-        This function is idempotent and will not raise an exception
-        if the function is not in the command's checks.
-
-        .. versionadded:: 1.3
-
-        Parameters
-        -----------
-        func
-            The function to remove from the checks.
-        """
-
-        try:
-            self.checks.remove(func)
-        except ValueError:
-            pass
-
-    def update(self, **kwargs: Any) -> None:
-        """Updates :class:`Command` instance with updated attribute.
-
-        This works similarly to the :func:`.command` decorator in terms
-        of parameters in that they are passed to the :class:`Command` or
-        subclass constructors, sans the name and callback.
-        """
-        self.__init__(self.callback, **
-                      dict(self.__original_kwargs__, **kwargs))
-
-    async def __call__(self, context: Context, *args: P.args, **kwargs: P.kwargs) -> T:
-        """|coro|
-
-        Calls the internal callback that the command holds.
-
-        .. note::
-
-            This bypasses all mechanisms -- including checks, converters,
-            invoke hooks, cooldowns, etc. You must take care to pass
-            the proper arguments and types to this function.
-
-        .. versionadded:: 1.3
-        """
-        if self.cog is not None:
-            # type: ignore
-            return await self.callback(self.cog, context, *args, **kwargs)
-        else:
-            # type: ignore
-            return await self.callback(context, *args, **kwargs)
-
-    def _ensure_assignment_on_copy(self, other: CommandT) -> CommandT:
-        other._before_invoke = self._before_invoke
-        other._after_invoke = self._after_invoke
-        if self.checks != other.checks:
-            other.checks = self.checks.copy()
-        if self._buckets.valid and not other._buckets.valid:
-            other._buckets = self._buckets.copy()
-        if self._max_concurrency != other._max_concurrency:
-            # _max_concurrency won't be None at this point
-            other._max_concurrency = self._max_concurrency.copy()  # type: ignore
-
-        try:
-            other.on_error = self.on_error
-        except AttributeError:
-            pass
-        return other
-
-    def copy(self: CommandT) -> CommandT:
-        """Creates a copy of this command.
-
-        Returns
-        --------
-        :class:`Command`
-            A new instance of this command.
-        """
-        ret = self.__class__(self.callback, **self.__original_kwargs__)
-        return self._ensure_assignment_on_copy(ret)
-
-    def _update_copy(self: CommandT, kwargs: Dict[str, Any]) -> CommandT:
-        if kwargs:
-            kw = kwargs.copy()
-            kw.update(self.__original_kwargs__)
-            copy = self.__class__(self.callback, **kw)
-            return self._ensure_assignment_on_copy(copy)
-        else:
-            return self.copy()
-
-    async def dispatch_error(self, ctx: Context, error: Exception) -> None:
-        ctx.command_failed = True
-        cog = self.cog
-        try:
-            coro = self.on_error
-        except AttributeError:
-            pass
-        else:
-            injected = wrap_callback(coro)
-            if cog is not None:
-                await injected(cog, ctx, error)
-            else:
-                await injected(ctx, error)
-
-        try:
-            if cog is not None:
-                local = Cog._get_overridden_method(cog.cog_command_error)
-                if local is not None:
-                    wrapped = wrap_callback(local)
-                    await wrapped(ctx, error)
-        finally:
-            ctx.bot.dispatch('command_error', ctx, error)
 
     async def transform(self, ctx: Context, param: inspect.Parameter) -> Any:
         required = param.default is param.empty
@@ -980,60 +712,6 @@ class Command(CommandBase, Generic[CogT, P, T]):
         .. versionadded:: 1.7
         """
         return hasattr(self, 'on_error')
-
-    def before_invoke(self, coro: HookT) -> HookT:
-        """A decorator that registers a coroutine as a pre-invoke hook.
-
-        A pre-invoke hook is called directly before the command is
-        called. This makes it a useful function to set up database
-        connections or any type of set up required.
-
-        This pre-invoke hook takes a sole parameter, a :class:`.Context`.
-
-        See :meth:`.Bot.before_invoke` for more info.
-
-        Parameters
-        -----------
-        coro: :ref:`coroutine <coroutine>`
-            The coroutine to register as the pre-invoke hook.
-
-        Raises
-        -------
-        TypeError
-            The coroutine passed is not actually a coroutine.
-        """
-        if not asyncio.iscoroutinefunction(coro):
-            raise TypeError('The pre-invoke hook must be a coroutine.')
-
-        self._before_invoke = coro
-        return coro
-
-    def after_invoke(self, coro: HookT) -> HookT:
-        """A decorator that registers a coroutine as a post-invoke hook.
-
-        A post-invoke hook is called directly after the command is
-        called. This makes it a useful function to clean-up database
-        connections or any type of clean up required.
-
-        This post-invoke hook takes a sole parameter, a :class:`.Context`.
-
-        See :meth:`.Bot.after_invoke` for more info.
-
-        Parameters
-        -----------
-        coro: :ref:`coroutine <coroutine>`
-            The coroutine to register as the post-invoke hook.
-
-        Raises
-        -------
-        TypeError
-            The coroutine passed is not actually a coroutine.
-        """
-        if not asyncio.iscoroutinefunction(coro):
-            raise TypeError('The post-invoke hook must be a coroutine.')
-
-        self._after_invoke = coro
-        return coro
 
     @property
     def cog_name(self) -> Optional[str]:
